Implement actions panel module
I wasn't completely satisfied with the styling of the built-in Angular
Material expansion panel actions, so I created an alternate "actions
panel" component which can be used for displaying actions. To avoid
congestion, it automatically moves all but the top three actions into an
overflow menu. It also allows for some additional configurability of
actions, such as utilizing a color hint and a 'prominence' value.